Got mE_client <-> up-linux-streamer <-> uE_service also working
