+++ up-linux-streamer/Cargo.toml
@@ -11,7 +11,7 @@ license.workspace = true
 [dependencies]
 tokio = { version = "1.35.1", default-features = false }
 up-rust = { git = "https://github.com/eclipse-uprotocol/up-rust", rev = "d736fdf35ff4728effa7f36b720f0fc1605d5ba0" }
-up-client-zenoh = { git = "https://github.com/ZettaScaleLabs/up-client-zenoh-rust", branch = "bump_1.5.8" }
+up-transport-zenoh = { git = "https://github.com/ZettaScaleLabs/up-client-zenoh-rust", branch = "bump_1.5.8" }
 up-transport-vsomeip = { git = "https://github.com/PLeVasseur/up-client-vsomeip-rust.git", branch = "feature/initial-implementation" }
 #up-transport-vsomeip = { git = "https://github.com/PLeVasseur/up-client-vsomeip-rust.git", branch = "feature/investigate_slice_from_raw_parts" }
 up-streamer = { path = "../up-streamer" }
