+++ up-linux-streamer/examples/mE_client.rs
@@ -1,17 +1,20 @@
 use async_trait::async_trait;
+use log::trace;
+use std::fs::canonicalize;
+use std::path::PathBuf;
 use std::sync::Arc;
 use std::time::Duration;
 use up_rust::{UListener, UMessage, UMessageBuilder, UPayloadFormat, UStatus, UTransport, UUri};
-use up_transport_zenoh::UPClientZenoh;
+use up_transport_vsomeip::UPTransportVsomeip;
 use zenoh::config::Config;
 
-const SERVICE_AUTHORITY: &str = "me_authority";
-const SERVICE_UE_ID: u16 = 0x4321;
+const SERVICE_AUTHORITY: &str = "linux";
+const SERVICE_UE_ID: u16 = 0x1236;
 const SERVICE_UE_VERSION_MAJOR: u8 = 1;
-const SERVICE_RESOURCE_ID: u16 = 0x0421;
+const SERVICE_RESOURCE_ID: u16 = 0x0896;
 
-const CLIENT_AUTHORITY: &str = "linux";
-const CLIENT_UE_ID: u16 = 0x1236;
+const CLIENT_AUTHORITY: &str = "me_authority";
+const CLIENT_UE_ID: u16 = 0x5678;
 const CLIENT_UE_VERSION_MAJOR: u8 = 1;
 const CLIENT_RESOURCE_ID: u16 = 0;
 
@@ -34,15 +37,23 @@ impl UListener for ServiceResponseListener {
 async fn main() -> Result<(), UStatus> {
     env_logger::init();
 
-    println!("uE_client");
+    println!("mE_client");
 
-    // TODO: Probably make somewhat configurable?
-    let zenoh_config = Config::default();
-    // TODO: Add error handling if we fail to create a UPClientZenoh
+    let crate_dir = env!("CARGO_MANIFEST_DIR");
+    // TODO: Make configurable to pass the path to the vsomeip config as a command line argument
+    let vsomeip_config = PathBuf::from(crate_dir).join("vsomeip-configs/mE_client.json");
+    let vsomeip_config = canonicalize(vsomeip_config).ok();
+    trace!("vsomeip_config: {vsomeip_config:?}");
+
+    // There will be a single vsomeip_transport, as there is a connection into device and a streamer
+    // TODO: Add error handling if we fail to create a UPTransportVsomeip
     let client: Arc<dyn UTransport> = Arc::new(
-        UPClientZenoh::new(zenoh_config, "linux".to_string())
-            .await
-            .unwrap(),
+        UPTransportVsomeip::new_with_config(
+            &CLIENT_AUTHORITY.to_string(),
+            CLIENT_UE_ID,
+            &vsomeip_config.unwrap(),
+        )
+        .unwrap(),
     );
 
     let source = UUri {
