+++ up-linux-streamer/vsomeip-configs/mE_client.json
@@ -1,0 +1,9 @@
+{
+  "applications" :
+  [
+    {
+      "name" : "0x5678",
+      "id" : "0x5678"
+    }
+  ]
+}
